+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [Unreleased]
 
+### Changed
+* Add project path and id to event logs
+
 ## [0.3.1] - 2026-03-18
 
 ### Fixed